Refactor: Started model layer pattern
