+++ 01_presentation/actions/actions.js
@@ -1,0 +1,9 @@
+const ArticleActions = require('./articleAction')
+
+class Actions {
+  static configure(express) {
+    this.articleActions = new ArticleActions(express)
+  }
+}
+
+module.exports = Actions
