+++ 02_domain/models/Article.js
