+++ 03_infra/api/cors.js
@@ -1,12 +1,8 @@
 const cors = require('cors')
 
 class Cors {
-  constructor() {
-    this.whiteList = process.env.ORIGINS
-  }
-
   allowOrigins(origin, callback) {
-    if (this.whiteList.indexOf(origin) !== -1) {
+    if (process.env.ORIGINS.indexOf(origin) !== -1 || !origin) {
       callback(null, true)
     } else {
       callback(new Error('Not allowed by CORS'))
