+++ 03_infra/database/repositories/baseRepository.js
@@ -1,0 +1,19 @@
+class BaseRepository {
+  constructor(model) {
+    this._model = model
+  }
+
+  get(skip = 0, take = 10, query = null) {
+    return this._model.find(query).skip(skip).limit(take)
+  }
+
+  getOne(id) {
+    return this._model.findById(id)
+  }
+
+  create(model) {
+    return new this._model(model).save()
+  }
+}
+
+module.exports = BaseRepository
