+++ 03_infra/database/unitOfWork.js
@@ -1,0 +1,9 @@
+const ArticleRepository = require('./repositories/articleRepository')
+
+class UnitOfWork {
+  constructor() {
+    this.articleRepository = new ArticleRepository()
+  }
+}
+
+module.exports = new UnitOfWork()
