+++ 03_infra/dependencyInjection/index.js
@@ -3,15 +3,14 @@ const consign = require('consign')
 class ServiceLocator {
   static configure(express) {
     consign()
-      .include('/config/validation.js')
+      .include('/03_infra/validation.js')
       .then('/api/managementHttpResponse.js')
       .then('/api/comments')
       .then('/api/views')
-      .then('/api/articles')
       .then('/api/likes')
       .then('/api/messages')
       .then('/api/themes')
-      .then('/api')
+      .then('/api/themes')
       .into(express)
   }
 }
