+++ 03_infra/services/index.js
@@ -1,6 +1,7 @@
 const DI = require('../dependencyInjection')
 const Database = require('../database')
 const Middlewares = require('../api/middlewares')
+const Actions = require('../../01_presentation/actions/actions')
 
 const Cors = require('../api/cors')
 
@@ -11,10 +12,15 @@ class Services {
     Middlewares.configure(express)
     Database.connect()
     DI.configure(express)
+    Actions.configure(express)
   }
 
   static configurePublicResources(express, path = '/public') {
     express().use(path, express.static('public'))
+  }
+
+  static configureActions(express) {
+
   }
 
   static start(express, port = 3010, host = '0.0.0.0') {
