+++ index.js
@@ -1,7 +1,8 @@
 const express = require('express')
 const app = express()
-const services = require('./config/services')
+const services = require('./03_infra/services')
 
 services.configurePublicResources(express)
 services.configure(app)
+services.configureActions(express)
 services.start(app)
